scripts: Add a scripts to create a bootable installation isos
This script takes a existing CentOS 7 iso and add's the given
squashfs and a minimalisitc script to generate a new and working
CentOS 7 iso for Node Next.

Usage:
$ bash derive-boot-iso.sh boot.iso ovirt-node-ng-image.squashfs.img

boot.iso and ovirt-node-ng-image.squashfs.img need to exist.

A convenience script is create-nightly-boot-iso.sh which will pull
a CentOS 7 installation ISO and the correct squashfs and generate
the bootable Node Next iso on the fly.

Usage:
$ bash create-nightly-boot-iso.sh

Change-Id: Ic3ed464a415118887777edf9b55427cd5ab4ca2d
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>