ks: Strip git-hash from image NVR
The image NVR is later used for the LV anem holding the image.
Thsi LV name has a limit of 55 chars, which is quickly approached if
we push so many informations (name, version, pre-release, release, build
date, git-hash, dist) into the NVR.

The git-hash is now dropped, because it uses a couple of chars, and is the
least useful information.

The complete informations can still be retrieved by using

$ rpm -q ovirt-release-host-node

Change-Id: Ie92e6310e949e622b000440355d308888c4a8d7d
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1322038
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>