spec: Add fu to align the image version and release
Previously the image and image-update version and release was
defined by this package.
More important, the version and release was different that the
version and release of the image-update-placeholder package inside
the image.
This lead to problems in the upgrade logic.

Now the version and release of the wrapper rpm containing the image
will be taken from the image-update-placeholder package inside
the image.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1334874
Change-Id: If6f87902cd75d0b8462df5d97d1eee8fb8983dc2
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>