tests: Improve event catching
It cna be that we miss an event, because we start listening for events
after we triggered an action (i.e. reboot) before we start watching for
the event (lifecycle).
With this patch we now start listening (in a thread) before we trigger
the action.
This should allow us to not miss an event.

This bug cna actualy lead to timeouts in tests if we miss an event we
are waiting for.

Change-Id: I2cfde443ea71fdb4f90cade5206b63d97406896c
Signed-off-by: Fabian Deutsch <fabiand@fedoraproject.org>